+++ apps/dsb-client-gateway-api-e2e/src/app/channel-controller.e2e-spec.ts
@@ -3,7 +3,7 @@ import { INestApplication } from '@nestjs/common';
 import request from 'supertest';
 import { AppModule } from '../../../dsb-client-gateway-api/src/app/app.module';
 import { DsbApiService } from '../../../dsb-client-gateway-api/src/app/modules/dsb-client/service/dsb-api.service';
-import { IamFactoryService } from '../../../../libs/dsb-client-gateway-iam-client/src/lib/service/iam-factory.service';
+import { IamFactoryService } from '@dsb-client-gateway/dsb-client-gateway-iam-client';
 import { CreateChannelDto } from '../../../dsb-client-gateway-api/src/app/modules/channel/dto/request/create-channel.dto';
 import { ChannelType } from '../../../dsb-client-gateway-api/src/app/modules/channel/channel.const';
 import { clearDatabase } from './utils/lokijs-utils';
