+++ apps/dsb-client-gateway-api/src/app/modules/identity/identity.module.ts
@@ -5,10 +5,11 @@ import { IdentityController } from './identity.controller';
 import { StorageModule } from '../storage/storage.module';
 import { EnrolmentModule } from '../enrolment/enrolment.module';
 import { SecretsEngineModule } from '@dsb-client-gateway/dsb-client-gateway-secrets-engine';
+import { IamInitService } from './service/iam-init.service';
 
 @Module({
   imports: [UtilsModule, StorageModule, SecretsEngineModule, EnrolmentModule],
-  providers: [IdentityService],
+  providers: [IdentityService, IamInitService],
   controllers: [IdentityController],
   exports: [IdentityService],
 })
