+++ apps/dsb-client-gateway-api/src/app/modules/identity/service/iam-init.service.ts
@@ -1,8 +1,8 @@
 import { Injectable, Logger, OnModuleInit } from '@nestjs/common';
 import { IamService } from '@dsb-client-gateway/dsb-client-gateway-iam-client';
-import { SecretsEngineService } from '../secrets-engine.interface';
-import { EnrolmentRepository } from '../../../../../apps/dsb-client-gateway-api/src/app/modules/storage/repository/enrolment.repository';
-import { IdentityRepository } from '../../../../../apps/dsb-client-gateway-api/src/app/modules/storage/repository/identity.repository';
+import { SecretsEngineService } from '@dsb-client-gateway/dsb-client-gateway-secrets-engine';
+import { EnrolmentRepository } from '../../storage/repository/enrolment.repository';
+import { IdentityRepository } from '../../storage/repository/identity.repository';
 
 @Injectable()
 export class IamInitService implements OnModuleInit {
