+++ apps/dsb-client-gateway-scheduler/src/app/modules/did/handler/index.ts
@@ -1,0 +1,1 @@
+export * from './did-attribute-changed.handler';
