+++ apps/dsb-client-gateway-scheduler/src/app/modules/did/index.ts
@@ -1,0 +1,3 @@
+export * from './service';
+export * from './handler';
+export * from './did.module';
