+++ apps/dsb-client-gateway-scheduler/src/app/modules/did/service/index.ts
@@ -1,0 +1,1 @@
+export * from './did-listener.service';
