+++ libs/dsb-client-gateway-did-registry/src/lib/command/index.ts
@@ -1,0 +1,1 @@
+export * from './did-attribute-changed.command';
