+++ libs/dsb-client-gateway-did-registry/src/lib/index.ts
@@ -1,2 +1,4 @@
 export * from './ethers';
 export * from './did-registry.module';
+export * from './service';
+export * from './command';
