+++ libs/dsb-client-gateway-did-registry/src/lib/service/index.ts
@@ -1,0 +1,2 @@
+export * from './did-registry-listener.service';
+export * from './provider';
