+++ libs/dsb-client-gateway-iam-client/src/lib/iam.interface.ts
@@ -1,0 +1,10 @@
+import { IAppDefinition } from '@energyweb/iam-contracts';
+
+export class ApplicationDTO implements IAppDefinition {
+  appName: string;
+  logoUrl?: string;
+  websiteUrl?: string;
+  description?: string;
+  namespace?: string;
+  topicsCount?: number;
+}
