+++ libs/dsb-client-gateway-iam-client/src/lib/index.ts
@@ -1,2 +1,3 @@
 export * from './iam.module';
 export * from './service';
+export * from './iam.interface';
