+++ libs/dsb-client-gateway-secrets-engine/src/lib/secrets-engine.module.ts
@@ -4,11 +4,8 @@ import { ConfigService } from '@nestjs/config';
 import { AwsSsmService } from './service/aws-ssm.service';
 import { SecretsEngineService } from './secrets-engine.interface';
 import { VaultService } from './service/vault.service';
-import { IamInitService } from './service/iam-init.service';
-import { StorageModule } from '../../../../apps/dsb-client-gateway-api/src/app/modules/storage/storage.module';
 
 @Module({
-  imports: [StorageModule],
   providers: [
     {
       provide: SecretsEngineService,
@@ -30,7 +27,6 @@ import { StorageModule } from '../../../../apps/dsb-client-gateway-api/src/app/m
       },
       inject: [ConfigService],
     },
-    IamInitService,
   ],
   exports: [SecretsEngineService],
 })
