+++ libs/dsb-client-gateway-secrets-engine/src/lib/service/index.ts
@@ -1,3 +1,2 @@
 export * from './aws-ssm.service';
-export * from './iam-init.service';
 export * from './vault.service';
