+++ libs/dsb-client-gateway-storage/src/lib/repository/index.ts
@@ -1,1 +1,2 @@
 export * from './abstract-loki.repository';
+export * from './did.repository';
